move out centroids
